Add missing override functions for logging functions
